# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Card Shoe API \u2014 proposed contract",
    "version": "0.1.0-draft",
    "description": "Endpoints and centralized exception handling are implemented. Request validation is implemented; additional HTTP error mappings remain planned. See implementation-plan.md for confirmed behavior and remaining proposals. Mock responses are stateless examples, not game behavior. JSON request objects reject unknown fields, missing/null fields and invalid types; no numeric coercion. Errors use detail, status, and a stable code. Missing resources return 404, already attached decks return 422, invalid bodies and UUID path parameters return 400, and other exceptions currently return 500; additional HTTP error responses below remain planned. Unknown routes return 404 ENDPOINT_NOT_FOUND. In-memory storage remains a proposal."
  },
  "servers": [
    {
      "url": "http://localhost:4010",
      "description": "Stateless contract mock (Prism; start separately)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Planned Spring API (endpoints not implemented yet)"
    }
  ],
...
  "paths": {
    "/api/v1/games": {
      "post": {
        "operationId": "createGame",
        "tags": [
          "Games"
        ],
        "summary": "Create an empty game",
        "description": "Create a named game with an empty shoe and no players. In-memory state is lost on restart.",
        "responses": {
          "201": {
            "description": "Create an empty game.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGameResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "id": "11111111-1111-4111-8111-111111111111",
                      "name": "Example game"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "detail": "Request body must be valid JSON.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  },
                  "INVALID_PARAMETER": {
                    "value": {
                      "detail": "Parameter 'gameId' must be a UUID.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  },
                  "INVALID_FIELDS": {
                    "value": {
                      "detail": "One or more request fields are invalid.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "METHOD_NOT_ALLOWED": {
                    "value": {
                      "detail": "The HTTP method is not supported for this endpoint.",
                      "status": 405,
                      "code": "METHOD_NOT_ALLOWED"
                    }
                  }
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Methods supported by this resource, including applicable HEAD/OPTIONS.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Client does not accept the response media type. The body may be empty if application/problem+json is also unacceptable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "NOT_ACCEPTABLE": {
                    "value": {
                      "detail": "The requested response media type is not supported.",
                      "status": 406,
                      "code": "NOT_ACCEPTABLE"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "status": 500,
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "UNSUPPORTED_MEDIA_TYPE": {
                    "value": {
                      "detail": "Content-Type must be application/json.",
                      "status": 415,
                      "code": "UNSUPPORTED_MEDIA_TYPE"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameRequest"
              },
              "example": {
                "name": "Example game"
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{gameId}": {
      "delete": {
        "operationId": "deleteGame",
        "tags": [
          "Games"
        ],
        "summary": "Delete a game",
        "description": "Delete the game and everything it owns: shoe, assigned decks, cards, players and hands. Unassigned decks survive. Players must be added again for a new game. Repeat deletion returns 404. No request body is accepted.",
        "responses": {
          "204": {
            "description": "No response body."
          },
          "400": {
            "description": "Bad Request. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "detail": "Request body must be valid JSON.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  },
                  "INVALID_PARAMETER": {
                    "value": {
                      "detail": "Parameter 'gameId' must be a UUID.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "METHOD_NOT_ALLOWED": {
                    "value": {
                      "detail": "The HTTP method is not supported for this endpoint.",
                      "status": 405,
                      "code": "METHOD_NOT_ALLOWED"
                    }
                  }
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Methods supported by this resource, including applicable HEAD/OPTIONS.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Client does not accept the response media type. The body may be empty if application/problem+json is also unacceptable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "NOT_ACCEPTABLE": {
                    "value": {
                      "detail": "The requested response media type is not supported.",
                      "status": 406,
                      "code": "NOT_ACCEPTABLE"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "status": 500,
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "GAME_NOT_FOUND": {
                    "value": {
                      "detail": "Game '11111111-1111-4111-8111-111111111111' was not found.",
                      "status": 404,
                      "code": "GAME_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
...
    "/api/v1/games/{gameId}/decks": {
      "post": {
        "operationId": "addDeck",
        "tags": [
          "Decks"
        ],
        "summary": "Permanently add a deck to a game",
        "description": "Attach an unassigned deck to this game shoe. A deck belongs to at most one shoe, and a shoe belongs to one game. Assignment is atomic; attached decks cannot be removed. Duplicate or cross-game attachment returns 422 DECK_ALREADY_ASSIGNED.",
        "responses": {
          "204": {
            "description": "Deck added. No response body."
          },
          "400": {
            "description": "Bad Request. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "detail": "Request body must be valid JSON.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  },
                  "INVALID_PARAMETER": {
                    "value": {
                      "detail": "Parameter 'gameId' must be a UUID.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  },
                  "INVALID_FIELDS": {
                    "value": {
                      "detail": "One or more request fields are invalid.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "METHOD_NOT_ALLOWED": {
                    "value": {
                      "detail": "The HTTP method is not supported for this endpoint.",
                      "status": 405,
                      "code": "METHOD_NOT_ALLOWED"
                    }
                  }
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Methods supported by this resource, including applicable HEAD/OPTIONS.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Client does not accept the response media type. The body may be empty if application/problem+json is also unacceptable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "NOT_ACCEPTABLE": {
                    "value": {
                      "detail": "The requested response media type is not supported.",
                      "status": 406,
                      "code": "NOT_ACCEPTABLE"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "status": 500,
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "UNSUPPORTED_MEDIA_TYPE": {
                    "value": {
                      "detail": "Content-Type must be application/json.",
                      "status": 415,
                      "code": "UNSUPPORTED_MEDIA_TYPE"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "GAME_NOT_FOUND": {
                    "value": {
                      "detail": "Game '11111111-1111-4111-8111-111111111111' was not found.",
                      "status": 404,
                      "code": "GAME_NOT_FOUND"
                    }
                  },
                  "DECK_NOT_FOUND": {
                    "value": {
                      "detail": "Deck '22222222-2222-4222-8222-222222222222' was not found.",
                      "status": 404,
                      "code": "DECK_NOT_FOUND"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Deck is already attached. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "DECK_ALREADY_ASSIGNED": {
                    "value": {
                      "detail": "Deck '22222222-2222-4222-8222-222222222222' is already assigned to a game.",
                      "status": 422,
                      "code": "DECK_ALREADY_ASSIGNED"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDeckRequest"
              },
              "example": {
                "deckId": "22222222-2222-4222-8222-222222222222"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
    "/api/v1/games/{gameId}/decks/cards": {
      "get": {
        "operationId": "countCards",
        "tags": [
          "Decks"
        ],
        "summary": "Count undealt copies of each face",
        "description": "Count undealt cards by suit and rank, including zeros. Return suit objects in hearts/spades/clubs/diamonds order, with ranks King down to Ace. Only undealt cards contribute. No request body is accepted.",
        "responses": {
          "200": {
            "description": "Count undealt copies of each face.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUndealtCardsResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "hearts": {
                        "king": 1,
                        "queen": 1,
                        "jack": 1,
                        "10": 1,
                        "9": 1,
                        "8": 1,
                        "7": 1,
                        "6": 1,
                        "5": 1,
                        "4": 1,
                        "3": 1,
                        "2": 1,
                        "ace": 1
                      },
                      "spades": {
                        "king": 1,
                        "queen": 1,
                        "jack": 1,
                        "10": 1,
                        "9": 1,
                        "8": 1,
                        "7": 1,
                        "6": 1,
                        "5": 1,
                        "4": 1,
                        "3": 1,
                        "2": 1,
                        "ace": 1
                      },
                      "clubs": {
                        "king": 1,
                        "queen": 1,
                        "jack": 1,
                        "10": 1,
                        "9": 1,
                        "8": 1,
                        "7": 1,
                        "6": 1,
                        "5": 1,
                        "4": 1,
                        "3": 1,
                        "2": 1,
                        "ace": 1
                      },
                      "diamonds": {
                        "king": 1,
                        "queen": 1,
                        "jack": 1,
                        "10": 1,
                        "9": 1,
                        "8": 1,
                        "7": 1,
                        "6": 1,
                        "5": 1,
                        "4": 1,
                        "3": 1,
                        "2": 1,
                        "ace": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "detail": "Request body must be valid JSON.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  },
                  "INVALID_PARAMETER": {
                    "value": {
                      "detail": "Parameter 'gameId' must be a UUID.",
                      "status": 400,
                      "code": "BAD_REQUEST"
                    }
                  }
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "METHOD_NOT_ALLOWED": {
                    "value": {
                      "detail": "The HTTP method is not supported for this endpoint.",
                      "status": 405,
                      "code": "METHOD_NOT_ALLOWED"
                    }
                  }
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Methods supported by this resource, including applicable HEAD/OPTIONS.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Client does not accept the response media type. The body may be empty if application/problem+json is also unacceptable.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "NOT_ACCEPTABLE": {
                    "value": {
                      "detail": "The requested response media type is not supported.",
                      "status": 406,
                      "code": "NOT_ACCEPTABLE"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "detail": "An unexpected error occurred.",
                      "status": 500,
                      "code": "INTERNAL_ERROR"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found. See named examples and shared error catalogue.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "GAME_NOT_FOUND": {
                    "value": {
                      "detail": "Game '11111111-1111-4111-8111-111111111111' was not found.",
                      "status": 404,
                      "code": "GAME_NOT_FOUND"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddDeckRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "deckId"
        ],
        "properties": {
          "deckId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateGameRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "\\S",
            "description": "Game name must contain at least one non-whitespace character."
          }
        }
      },
...
      "CreateGameResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "detail",
          "status",
          "code"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          }
        }
      },
...
      "GetUndealtCardsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "hearts",
          "spades",
          "clubs",
          "diamonds"
        ],
        "properties": {
          "hearts": {
            "type": "object",
            "$ref": "#/components/schemas/GetUndealtSuitCardsResponse"
          },
          "spades": {
            "type": "object",
            "$ref": "#/components/schemas/GetUndealtSuitCardsResponse"
          },
          "clubs": {
            "type": "object",
            "$ref": "#/components/schemas/GetUndealtSuitCardsResponse"
          },
          "diamonds": {
            "type": "object",
            "$ref": "#/components/schemas/GetUndealtSuitCardsResponse"
          }
        }
      },
      "GetUndealtSuitCardsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "king",
          "queen",
          "jack",
          "10",
          "9",
          "8",
          "7",
          "6",
          "5",
          "4",
          "3",
          "2",
          "ace"
        ],
        "properties": {
          "king": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "queen": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "jack": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "10": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "9": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "8": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "7": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "6": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "5": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "4": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "3": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "2": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "ace": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    }
  }
}